Changed how I render hobby images
After attempt after attempt, I was unable to get the background image fixed on scroll for mobile in production.  It worked fine on desktop view, but not mobile.  In development, it worked on all screens, but this was pre-build so I'm not surpised.  I couldn't find any satisfactory solutions yet and I do not want to waste too much time on this.